diabeticRetinopathy/requirements.txt
Flask
gunicorn
tensorflow
tensorflow_hub
numpy
pillow
